Refactorinf test codes for forecast views
